Additional Clean up
